[MOVE] 일정 관련 파일 이동
