+++ Sofa/Sources/View/Calendar/SubViews/Task/TaskColorPicker.swift
